+++ src/views/Partners.js
@@ -1,48 +1,110 @@
-import angostiqLogo from './../images/agnostiq_logo.png'
-import cqcLogo from './../images/cqc_logo.png'
-import iqtLogo from './../images/iqt_logo.png'
-import rikenLogo from './../images/riken_logo.png'
-import ufLogo from './../images/unitary_fund_logo.png'
-import usraLogo from './../images/usra_logo.png'
-import strangeworksLogo from './../images/strangeworks_logo.png'
-import quantonationLogo from './../images/quantonation_logo.png'
-import sqmsLogo from './../images/sqms_logo.png'
-import supertechLogo from './../images/supertech_logo.png'
-import quantumcomputingreportLogo from './../images/quantumcomputingreport_logo.png'
-import bluequbitLogo from './../images/bluequbit_logo.png'
+import axios from 'axios'
+import React from 'react'
+import { sortCommon, sortAlphabetical } from '../components/SortFunctions'
+import { withRouter, Link } from 'react-router-dom'
+import { library } from '@fortawesome/fontawesome-svg-core'
+import { faHeart, faExternalLinkAlt, faChartLine } from '@fortawesome/free-solid-svg-icons'
+import config from '../config'
+import ErrorHandler from '../components/ErrorHandler'
+import Partners from './Partners'
+import Committee from './Committee'
 
-const Partners = () => {
-  return (
-    <div className='row'>
-      <div className='col-md-2' />
-      <div className='col-md-8 text-justify'>
-        <h2 className='text-center'>Our Partners and Supporters</h2>
-        <div>
-          <p>
-            We thank the following organizations and institutions for their support.
-          </p>
-          <p>
-            If your organization or institution would like to become a partner, please contact us at <a href='mailto:metriq@unitary.fund'>metriq@unitary.fund</a>.
-          </p>
-          <div className='text-center'>
-            <a href='https://www.iqt.org/'><img src={iqtLogo} alt='IQT Labs partner logo' className='logo-image' /></a>
-            <a href='https://unitary.fund/'><img src={ufLogo} alt='UF partner logo' className='logo-image' /></a>
-            <a href='https://riacs.usra.edu/quantum/nisqc-nl'><img src={usraLogo} alt='USRA partner logo' className='logo-image' /></a>
-            <a href='https://www.bluequbit.io/'><img src={bluequbitLogo} alt='BlueQubit logo' className='logo-image' /></a><br />
-            <a href='https://cambridgequantum.com/'><img src={cqcLogo} alt='Cambridge Quantum Computing partner logo' className='logo-image' /></a>
-            <a href='https://www.riken.jp/en/'><img src={rikenLogo} alt='Riken University logo' className='logo-image' /></a>
-            <a href='https://strangeworks.com/'><img src={strangeworksLogo} alt='Strangeworks logo' className='logo-image' /></a>
-            <a href='https://sqms.fnal.gov/'><img src={sqmsLogo} alt='SQMS logo' className='logo-image' /></a><br />
-            <a href='https://www.quantonation.com/'><img src={quantonationLogo} alt='Quantonation logo' className='logo-image' /></a>
-            <a href='https://agnostiq.ai/'><img src={angostiqLogo} alt='Agnostiq logo' className='logo-image' /></a>
-            <a href='https://super.tech/'><img src={supertechLogo} alt='Super.Tech logo' className='logo-image' /></a>
-            <a href='https://quantumcomputingreport.com/'><img src={quantumcomputingreportLogo} alt='Quantum Computing Report logo' className='logo-image' /></a>
-          </div>
+library.add(faHeart, faExternalLinkAlt, faChartLine)
+
+class Home extends React.Component {
+  constructor (props) {
+    super(props)
+    this.state = {
+      isLoading: true,
+      alphabetical: [],
+      allNames: [],
+      platforms: [],
+      trending: [],
+      popular: [],
+      latest: [],
+      topSubmitters: [],
+      activeTab: 'Trending',
+      filterId: null,
+      requestFailedMessage: '',
+      isLinkBlocked: false
+    }
+
+    this.handleOnFilter = this.handleOnFilter.bind(this)
+    this.handleOnSelect = this.handleOnSelect.bind(this)
+  }
+
+  handleOnFilter (value) {
+    if (value) {
+      this.setState({ filterId: value.id })
+    }
+  }
+
+  handleOnSelect (value) {
+    if (value) {
+      this.props.history.push('/Task/' + value.id)
+    }
+  }
+
+  componentDidMount () {
+    axios.get(config.api.getUriPrefix() + '/task/submissionCount')
+      .then(res => {
+        const alphabetical = res.data.data
+        alphabetical.sort(sortAlphabetical)
+        this.setState({ alphabetical, isLoading: false })
+      })
+      .catch(err => {
+        this.setState({ requestFailedMessage: ErrorHandler(err) })
+      })
+
+    axios.get(config.api.getUriPrefix() + '/task/names')
+      .then(res => {
+        this.setState({
+          requestFailedMessage: '',
+          allNames: res.data.data
+        })
+      })
+      .catch(err => {
+        this.setState({ requestFailedMessage: ErrorHandler(err) })
+      })
+
+    axios.get(config.api.getUriPrefix() + '/platform/submissionCount')
+      .then(res => {
+        const common = [...res.data.data]
+        common.sort(sortCommon)
+        const rws = []
+        for (let i = 0; i < 2; ++i) {
+          const row = []
+          for (let j = 0; j < 3; ++j) {
+            row.push(common[3 * i + j])
+          }
+          rws.push(row)
+        }
+        this.setState({
+          requestFailedMessage: '',
+          platforms: rws
+        })
+      })
+      .catch(err => {
+        this.setState({ requestFailedMessage: ErrorHandler(err) })
+      })
+  }
+
+  render () {
+    return (
+      <span>
+        <div className='metriq-header'>
+          <h1>Open tracking of quantum technology progress</h1>
+          <br />
         </div>
-      </div>
-      <div className='col-md-2' />
-    </div>
-  )
+        <div id='metriq-main-content'>
+          <p className='text-start'>Metriq is a platform for tracking and sharing quantum technology benchmarks. Users can make new <Link to='/Submissions'>submissions</Link> that show the performance of different <Link to='/Methods'>methods</Link> on <Link to='/Platforms'>platforms</Link> against <Link to='/Tasks'>tasks</Link>.</p>
+          <br />
+          <Partners />
+          <Committee />
+        </div>
+      </span>
+    )
+  }
 }
 
-export default Partners
+export default withRouter(Home)
